update : still resolving conflict between next.js and react
