+++ client/app/api/auth/[...nextauth]/route.js
@@ -1,3 +1,3 @@
 // Here the GET and POST is being imported from the auth.js file before we export fron here 
 
-export { GET, POST } from "@/app/auth/auth"
+export { GET, POST } from "@/app/auth/auth";
