+++ client/package.json
@@ -23,6 +23,7 @@
     "date-fns": "^3.6.0",
     "lucide-react": "^0.511.0",
     "next": "13.4.9",
+    "next-auth": "^4.24.11",
     "postcss": "8.4.25",
     "react": "18.2.0",
     "react-day-picker": "^8.10.1",
